+++ dts/save-structure/game-objects/game-object-behavior/known-behaviors/minion-identity.d.ts
@@ -1,3 +1,5 @@
+import { AccessoryType } from "../../../const-data";
+import { HashedString } from "../../../data-types";
 import { GameObjectBehavior } from "../game-object-behavior";
 import { BehaviorName } from "./types";
 export declare const MinionIdentityBehavior: BehaviorName<MinionIdentityBehavior>;
@@ -10,40 +12,20 @@ export interface MinionIdentityBehavior extends GameObjectBehavior {
         genderStringKey: MinionGender;
         arrivalTime: number;
         voiceIdx: number;
-        bodyData: {
-            headShape: {
-                hash: number;
-            };
-            mouth: {
-                hash: number;
-            };
-            neck: {
-                hash: number;
-            };
-            eyes: {
-                hash: number;
-            };
-            hair: {
-                hash: number;
-            };
-            body: {
-                hash: number;
-            };
-            arms: {
-                hash: number;
-            };
-            hat: {
-                hash: number;
-            };
-            hatHair: {
-                hash: number;
-            };
-            hairAlways: {
-                hash: number;
-            };
-        };
+        bodyData: BodyData;
     };
+}
+export interface BodyData {
+    headShape: HashedString;
+    mouth: HashedString;
+    neck: HashedString;
+    eyes: HashedString;
+    hair: HashedString;
+    body: HashedString;
+    arms: HashedString;
+    hat: HashedString;
 }
 export declare type MinionGender = "MALE" | "FEMALE" | "NB";
 export declare const MINION_IDENTITY_GENDERS: MinionGender[];
 export declare const MINION_IDENTITY_VOICES: number[];
+export declare const MINION_IDENTITY_BODY_DATA_ACCESSORIES: Record<keyof BodyData, AccessoryType>;