+++ src/save-structure/game-objects/game-object-behavior/known-behaviors/minion-identity.ts
@@ -1,4 +1,8 @@
+import { AccessoryType } from "../../../const-data";
+import { HashedString } from "../../../data-types";
+
 import { GameObjectBehavior } from "../game-object-behavior";
+
 import { BehaviorName } from "./types";
 
 export const MinionIdentityBehavior: BehaviorName<MinionIdentityBehavior> =
@@ -16,39 +20,19 @@ export interface MinionIdentityBehavior extends GameObjectBehavior {
 
     voiceIdx: number;
 
-    bodyData: {
-      headShape: {
-        hash: number;
-      };
-      mouth: {
-        hash: number;
-      };
-      neck: {
-        hash: number;
-      };
-      eyes: {
-        hash: number;
-      };
-      hair: {
-        hash: number;
-      };
-      body: {
-        hash: number;
-      };
-      arms: {
-        hash: number;
-      };
-      hat: {
-        hash: number;
-      };
-      hatHair: {
-        hash: number;
-      };
-      hairAlways: {
-        hash: number;
-      };
-    };
+    bodyData: BodyData;
   };
+}
+
+export interface BodyData {
+  headShape: HashedString;
+  mouth: HashedString;
+  neck: HashedString;
+  eyes: HashedString;
+  hair: HashedString;
+  body: HashedString;
+  arms: HashedString;
+  hat: HashedString;
 }
 
 export type MinionGender = "MALE" | "FEMALE" | "NB";
@@ -56,3 +40,17 @@ export type MinionGender = "MALE" | "FEMALE" | "NB";
 export const MINION_IDENTITY_GENDERS: MinionGender[] = ["MALE", "FEMALE", "NB"];
 
 export const MINION_IDENTITY_VOICES: number[] = [0, 1, 2, 3, 4];
+
+export const MINION_IDENTITY_BODY_DATA_ACCESSORIES: Record<
+  keyof BodyData,
+  AccessoryType
+> = {
+  headShape: "headshape",
+  mouth: "mouth",
+  neck: "neck",
+  eyes: "eyes",
+  hair: "hair",
+  body: "body",
+  arms: "arm",
+  hat: "hat"
+};