+++ src/types.d.ts
@@ -1,0 +1,1 @@
+declare type ArrayValues<T> = T extends (infer U)[] ? U : never;